Add a Jupyter notebook configuration file
Disable cross-site-request-forgery protection in the file to allow
Chrome and Safari browsers use the export to PDF feature.

The configuration file has the default file name. There is no need to
use a jupyter-lab flag or install the file as long as a jupyter-lab
command is executed in the same directory as the configuration file is
stored in.