Correctly use the shared buffer to transfer data
There are three buffers involved in IPC between two threads: a read
buffer, a write buffer and a shared buffer. The write buffer is the
source of data to be transferred, the read buffer is the storage for
data to be received, and the shared buffer should be used as an
intermediate buffer to transfer data between the sender and receiver.